% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{best_friend}
\alias{best_friend}
\title{Get the best friend of Mario by version.}
\usage{
best_friend(version)
}
\arguments{
\item{version}{Integer (1 or 2)}
}
\value{
Name of best friend (string)
}
\description{
Get the best friend of Mario by version.
}
\examples{
best_friend(1)
best_friend(2)
}
