% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_to_json.R
\name{file_to_json}
\alias{file_to_json}
\title{Turn a file that ends with a pipeline into JSON}
\usage{
file_to_json(path)
}
\arguments{
\item{path}{The path to an R code file.}
}
\description{
Turn a file that ends with a pipeline into JSON
}
