% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_args.R
\name{get_args}
\alias{get_args}
\title{Get the argument names and values from each step of the pipeline.}
\usage{
get_args(verbs)
}
\arguments{
\item{verbs}{A \code{\link[base:list]{base::list()}} of verbs from \code{\link[=get_verbs]{get_verbs()}}.}
}
\description{
Get the argument names and values from each step of the pipeline.
}
