% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_steps.R
\name{get_data_steps}
\alias{get_data_steps}
\title{See the data at each step of the pipeline}
\usage{
get_data_steps(verbs, envir = parent.frame())
}
\arguments{
\item{verbs}{A \code{\link[base:list]{base::list()}} of verbs from \code{\link[=get_verbs]{get_verbs()}}.}

\item{envir}{An \code{\link[base:environment]{base::environment()}} where the pipeline will be evaluated.}
}
\description{
See the data at each step of the pipeline
}
