% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\title{Get the data and function names from each step of the pipeline.}
\usage{
get_names(verbs)
}
\arguments{
\item{verbs}{A \code{\link[base:list]{base::list()}} of verbs from \code{\link[=get_verbs]{get_verbs()}}.}
}
\description{
Get the data and function names from each step of the pipeline.
}
