% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_verbs.R
\name{get_verbs}
\alias{get_verbs}
\title{Get the verbs from a pipeline}
\usage{
get_verbs(call)
}
\arguments{
\item{call}{A pipeline call, likely from either \code{\link[=parse_pipeline]{parse_pipeline()}}
or \code{\link[rlang:parse_expr]{rlang::parse_expr()}}.}
}
\value{
A \code{\link[base:list]{base::list()}} of \link[base:name]{base::name}s or \code{\link[base:call]{base::call()}}s.
}
\description{
Get the verbs from a pipeline
}
\examples{
library(dplyr)

mtcars \%>\%
  slice(1:2) \%>\%
  parse_pipeline() \%>\%
  get_verbs()

#>
}
