% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_functions.R
\name{mario_auth}
\alias{mario_auth}
\alias{mario_api_key}
\title{Authorization Header for mario}
\usage{
mario_auth(api_key = Sys.getenv("CONNECT_API_KEY"))

mario_api_key(api_key = Sys.getenv("CONNECT_API_KEY"))
}
\arguments{
\item{api_key}{API key for RStudio Cnnect}
}
\value{
A \code{header} output
}
\description{
Authorization Header for mario
}
\examples{
mario_auth()
}
