% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_functions.R
\name{mario_content}
\alias{mario_content}
\alias{mario_write_video}
\alias{mario_process_images}
\alias{mario_subtitles}
\alias{open_video}
\title{Extract content from Mario output}
\usage{
mario_content(response, output_file = NULL)

mario_write_video(response, output_file = NULL)

mario_process_images(response)

mario_subtitles(response)

open_video(response, open = TRUE, output_file = NULL)
}
\arguments{
\item{response}{A \code{response} object, usually an output from
\code{mario} or \code{mario_translate}}

\item{output_file}{A file path to where the output video should be saved.}

\item{open}{should the video be opened on the local machine?}
}
\value{
A list/\code{data.frame} of output
}
\description{
Extract content from Mario output
}
