% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales.R
\name{mario_plot_sales}
\alias{mario_plot_sales}
\title{Plot sales of Mario games.}
\usage{
mario_plot_sales(type = "bar")
}
\arguments{
\item{type}{Type of plot ("bar" or "pie").}
}
\value{
ggplot of sales
}
\description{
Plot sales of Mario games.
}
\examples{
mario_plot_sales("bar")
mario_plot_sales("pie")
}
