% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mario_helpers.R
\name{mario_voices}
\alias{mario_voices}
\title{List Voices available}
\usage{
mario_voices(
  service = NULL,
  api_url = mario_api_url(),
  api_key = Sys.getenv("CONNECT_API_KEY"),
  ...
)
}
\arguments{
\item{service}{Service to use or list voices}

\item{api_url}{URL for the Plumber API}

\item{api_key}{API key for RStudio Cnnect}

\item{...}{additional options to send to \code{\link{GET}}}
}
\value{
A \code{data.frame} of voices
}
\description{
List Voices available
}
\examples{
if (mario_have_api_key()) {
  result <- mario_voices()
  utils::head(result)
}
}
