% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pipeline.R
\name{parse_pipeline}
\alias{parse_pipeline}
\title{Capture a pipeline as a call}
\usage{
parse_pipeline(pipeline)
}
\arguments{
\item{pipeline}{A data pipeline.}
}
\description{
Capture a pipeline as a call
}
