% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pipeline_tbl.R
\name{parse_pipeline_tbl}
\alias{parse_pipeline_tbl}
\title{Get lots of information about a pipeline as a data frame.}
\usage{
parse_pipeline_tbl(pipeline)
}
\arguments{
\item{pipeline}{A data pipeline.}
}
\description{
Get lots of information about a pipeline as a data frame.
}
