% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pipeline_tbl.R
\name{pipeline_tbl}
\alias{pipeline_tbl}
\title{Get lots of information about a pipeline as a data frame.}
\usage{
pipeline_tbl(pipeline_call)
}
\arguments{
\item{pipeline_call}{A \code{\link[base:call]{base::call()}} representing a pipeline.}
}
\description{
Get lots of information about a pipeline as a data frame.
}
