% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_to_json.R
\name{pipeline_to_json}
\alias{pipeline_to_json}
\title{Turn a pipeline call into JSON}
\usage{
pipeline_to_json(call)
}
\arguments{
\item{call}{A pipeline call, likely from either \code{\link[=parse_pipeline]{parse_pipeline()}}
or \code{\link[rlang:parse_expr]{rlang::parse_expr()}}.}
}
\description{
Turn a pipeline call into JSON
}
