% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales-data.R
\docType{data}
\name{sales}
\alias{sales}
\title{Sales data for selected Mario games}
\format{
A data frame with 7 rows and 4 variables:
\describe{
\item{Game}{name of the game}
\item{System}{game platform}
\item{Release}{year of game release}
\item{Sales}{games sold, in million units}
}
}
\source{
\url{https://www.gamecubicle.com/features-mario-units_sold_sales.htm}
}
\usage{
sales
}
\description{
A dataset containing the release dates, systems and sales of a few Mario games.
}
\keyword{datasets}
