% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_to_json.R
\name{string_to_json}
\alias{string_to_json}
\title{Turn a string of a pipeline into a JSON}
\usage{
string_to_json(code)
}
\arguments{
\item{code}{A string of R code.}
}
\description{
Turn a string of a pipeline into a JSON
}
