% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_diff.R
\name{tibble_diff}
\alias{tibble_diff}
\title{Find differences between two tibbles.}
\usage{
tibble_diff(x, y)
}
\arguments{
\item{x}{A \code{\link[tibble:tibble]{tibble::tibble()}}.}

\item{y}{A \code{\link[tibble:tibble]{tibble::tibble()}}.}
}
\description{
Find differences between two tibbles.
}
