% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{best_friend}
\alias{best_friend}
\title{Plot sales of Mario games. WABBALUBBADUBDUB}
\usage{
best_friend(version)
}
\arguments{
\item{type}{Type of plot ("bar" or "pie").}
}
\value{
ggplot of sales
}
\description{
Plot sales of Mario games. WABBALUBBADUBDUB
}
\examples{
mario_plot_sales("bar")
mario_plot_sales("pie")
}
