% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_steps.R
\name{get_data_steps}
\alias{get_data_steps}
\title{See the data at each step of the pipeline}
\usage{
get_data_steps(call, envir = parent.frame())
}
\arguments{
\item{call}{A pipeline call, likely from either \code{\link[=parse_pipeline]{parse_pipeline()}}
or \code{\link[rlang:parse_expr]{rlang::parse_expr()}}.}

\item{envir}{An \code{\link[base:environment]{base::environment()}} where the pipeline will be evaluated.}
}
\description{
See the data at each step of the pipeline
}
